requirements.txt
binance_futures_connector==4.0.0
hexbytes>=0.3.1
matplotlib==3.8.4
numerize==0.12
pybit==5.6.2
Pypubsub==4.0.3
python-dotenv==1.0.1
python_binance==1.0.17
python_okx==0.3.1
Requests==2.32.3
scikit_learn==1.5.1
setuptools==68.2.2
web3>=7.0.0
gmx_python_sdk @ git+https://github.com/50shadesofgwei/gmx_python_sdk_custom.git@main
